backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
greenlet==3.2.2
mysqlclient==2.2.7
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.2
SQLAlchemy==2.0.40
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2